Clean CostBreakdown Container
