+++ src/common/__tests__/formatters.js
@@ -1,5 +1,5 @@
 import { shallow } from 'enzyme';
-import { noNeg, capitalizeFirstLetter, formatBytes, formatPrice } from '../formatters';
+import { noNeg, capitalizeFirstLetter, formatBytes, formatPrice, transformProducts } from '../formatters';
 
 describe('Formatters', () => {
 
@@ -63,6 +63,100 @@ describe('Formatters', () => {
       const spans = output.find("span");
       expect(spans.length).toBe(2);
       expect(spans.first().props().children[1]).toBe(formattedValue);
+    });
+
+  });
+
+  describe('tTansformProducts', () => {
+    const days = {
+      day: {
+        day1: 42,
+        day2: 21
+      }
+    };
+
+    const costsByProductPerDay = {
+      product: {
+        product1: {...days},
+        product2: {...days}
+      }
+    };
+
+    const costsAll = {...days};
+
+    const costsMissingDays = {
+      product: {
+        product1: {...days},
+        product2: {
+          day: {
+            ...days.day,
+            day3: 84
+          }
+        },
+      }
+    };
+
+    const costsMissingKeys = {
+      product: {
+        ...costsByProductPerDay.product,
+        "": days
+      }
+    };
+
+    it('returns an empty array when invalid filter', () => {
+      expect(transformProducts(costsByProductPerDay, "region", "day")).toEqual([]);
+    });
+
+    it('returns an empty array when valid filter and invalid interval', () => {
+      expect(transformProducts(costsByProductPerDay, "product", "month")).toEqual([]);
+    });
+
+    it('returns an empty array when filter is "all" and invalid interval', () => {
+      expect(transformProducts(costsAll, "all", "month")).toEqual([]);
+    });
+
+    it('returns formatted array when valid filter and valid interval', () => {
+      const output = [{
+        key: "product1",
+        values: [["day1", days.day.day1], ["day2", days.day.day2]]
+      },{
+        key: "product2",
+        values: [["day1", days.day.day1], ["day2", days.day.day2]]
+      }];
+      expect(transformProducts(costsByProductPerDay, "product", "day")).toEqual(output);
+    });
+
+    it('returns formatted array when filter is "all" and valid interval', () => {
+      const output = [{
+        key: "Total",
+        values: [["day1", days.day.day1], ["day2", days.day.day2]]
+      }];
+      expect(transformProducts(costsAll, "all", "day")).toEqual(output);
+    });
+
+    it('fills missing days', () => {
+      const output = [{
+        key: "product1",
+        values: [["day1", days.day.day1], ["day2", days.day.day2], ["day3", 0]]
+      },{
+        key: "product2",
+        values: [["day1", days.day.day1], ["day2", days.day.day2], ["day3", costsMissingDays.product.product2.day.day3]]
+      }];
+      expect(transformProducts(costsMissingDays, "product", "day")).toEqual(output);
+    });
+
+    it('fills missing keys', () => {
+      const output = [{
+        key: "product1",
+        values: [["day1", days.day.day1], ["day2", days.day.day2]]
+      },{
+        key: "product2",
+        values: [["day1", days.day.day1], ["day2", days.day.day2]]
+      },{
+        key: "No product",
+        values: [["day1", days.day.day1], ["day2", days.day.day2]]
+      }];
+      expect(transformProducts(costsMissingKeys, "product", "day")).toEqual(output);
     });
 
   });
