+++ src/components/aws/costBreakdown/index.js
@@ -1,0 +1,5 @@
+import Chart from './Chart';
+
+export default {
+  Chart
+};
