+++ src/components/aws/index.js
@@ -1,7 +1,9 @@
 import Accounts from './accounts';
 import S3Analytics from './s3Analytics';
+import CostBreakdown from './costBreakdown';
 
 export default {
   Accounts,
-  S3Analytics
+  S3Analytics,
+  CostBreakdown
 };
