+++ src/containers/aws/CostBreakdownContainer.js
@@ -1,18 +1,14 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import NVD3Chart from 'react-nvd3';
-import * as d3 from 'd3';
-import Components from "../../components";
-
-import Actions from "../../actions";
-
+import Components from '../../components';
+import Actions from '../../actions';
 import s3square from '../../assets/s3-square.png';
-import 'nvd3/build/nv.d3.min.css';
 
 const TimerangeSelector = Components.Misc.TimerangeSelector;
 const Selector = Components.Misc.Selector;
 const Panel = Components.Misc.Panel;
+const Chart = Components.AWS.CostBreakdown.Chart;
 
 export class CostBreakdownContainer extends Component {
 
@@ -28,31 +24,6 @@ export class CostBreakdownContainer extends Component {
       nextProps.getCosts(nextProps.costsDates.startDate, nextProps.costsDates.endDate, [nextProps.costsFilter, nextProps.costsInterval]);
   }
 
-  transformProducts = (data, filter, interval) => {
-    if (filter === "all" && data.hasOwnProperty(interval))
-      return [{
-        key: "Total",
-        values: Object.keys(data[interval]).map((date) => ([date, data[interval][date]]))
-      }];
-    else if (!data.hasOwnProperty(filter))
-      return [];
-    let dates = [];
-    try {
-      Object.keys(data[filter]).forEach((key) => {
-        Object.keys(data[filter][key][interval]).forEach((date) => {
-          if (dates.indexOf(date) === -1)
-            dates.push(date);
-        })
-      });
-      return Object.keys(data[filter]).map((key) => ({
-        key: (key.length ? key : `No ${filter}`),
-        values: dates.map((date) => ([date, data[filter][key][interval][date] || 0]))
-      }));
-    } catch (e) {
-      return [];
-    }
-  };
-
   render() {
 
     const filters = {
@@ -62,55 +33,10 @@ export class CostBreakdownContainer extends Component {
       region: "Region"
     };
 
-    /* istanbul ignore next */
-    const context = {
-      formatXAxis: (d) => (d3.time.format('%x')(new Date(d))),
-      formatYAxis: (d) => ('$' + d3.format(',.2f')(d)),
-    };
-
-    const datum = (this.props.costsValues && this.props.costsInterval && this.props.costsFilter ? this.transformProducts(this.props.costsValues, this.props.costsFilter, this.props.costsInterval) : null);
-
-    const chart = (datum ? (
-      <NVD3Chart
-        id="barChart"
-        type="multiBarChart"
-        datum={datum}
-        context={context}
-        xAxis={{
-          tickFormat: {
-            name:'formatXAxis',
-            type:'function',
-          }
-        }}
-        yAxis={{
-          tickFormat: {
-            name:'formatYAxis',
-            type:'function',
-          }
-        }}
-        margin={{right:100}}
-        rightAlignYAxis={true}
-        clipEdge={true}
-        showControls={true}
-        stacked={true}
-        x={
-          /* istanbul ignore next */
-          (d) => {
-          const date = new Date(d[0]);
-          return date.getTime();
-        }}
-        y={
-          /* istanbul ignore next */
-          (d) => (d[1])
-        }
-        height={600}
-      />
-    ) : null);
-
     return(
       <Panel>
 
-        <div>
+        <div className="clearfix">
           <h3 className="white-box-title no-padding inline-block">
             <img className="white-box-title-icon" src={s3square} alt="AWS square logo"/>
             Cost Breakdown
@@ -135,12 +61,10 @@ export class CostBreakdownContainer extends Component {
             </div>
           </div>
 
-          <div className="clearfix"/>
-
         </div>
 
         <div>
-          {chart}
+          <Chart values={this.props.costsValues} interval={this.props.costsInterval} filter={this.props.costsFilter}/>
         </div>
 
       </Panel>
@@ -154,8 +78,8 @@ CostBreakdownContainer.propTypes = {
     startDate: PropTypes.object,
     endDate: PropTypes.object,
   }),
-  costsInterval:PropTypes.string.isRequired,
-  costsFilter:PropTypes.string.isRequired,
+  costsInterval: PropTypes.string.isRequired,
+  costsFilter: PropTypes.string.isRequired,
   getCosts: PropTypes.func.isRequired,
   setCostsDates: PropTypes.func.isRequired,
   setCostsInterval: PropTypes.func.isRequired,
